Addintion of multiple files
Added given files AOloop_nocontrol.m and data .mat files.
Created Hankel.m, which builds a Hankel matrix of a given signal
Created Cphi.m, which computes the covariance matrix C(tau) of a given signal (for this assignment phi).
Created computeKalmanAR.m, which is to be written as part of Ex 2.6 to compute Kalmain gain-matrix and other matrices.
Created FinalAssignment.m where the whole assignment will be run from